style: 修改GroupModals组件为规范的三文件形式
