+++ src/components/Group/GroupModals/CreateGroupModal/index.type.ts
@@ -1,0 +1,5 @@
+export interface CreateGroupModalProps {
+  open: boolean;
+  onCancel: () => void;
+  onSuccess?: () => void;
+}
