+++ src/components/Group/GroupModals/DissolveGroupModal/index.module.less
@@ -1,0 +1,11 @@
+.modalSection {
+  margin-top: var(--ant-margin-sm);
+}
+
+.modalSectionLabel {
+  margin-bottom: var(--ant-margin-xs);
+}
+
+.modalSectionSubLabel {
+  color: var(--ant-color-text-secondary);
+}
