+++ src/components/Group/GroupModals/DissolveGroupModal/index.type.ts
@@ -1,0 +1,7 @@
+export interface DissolveGroupModalProps {
+  open: boolean;
+  onCancel: () => void;
+  onSuccess?: () => void;
+  groupId: string;
+  groupName: string;
+}
