+++ src/components/Group/GroupModals/EditGroupInfoModal/index.module.less
@@ -1,0 +1,3 @@
+.modalFormPadding {
+  padding: var(--ant-padding-lg) 0;
+}
