+++ src/components/Group/GroupModals/EditGroupInfoModal/index.type.ts
@@ -1,0 +1,10 @@
+export interface EditGroupInfoModalProps {
+  open: boolean;
+  onCancel: () => void;
+  onSuccess?: () => void;
+  groupId?: string;
+  groupName?: string;
+  description?: string;
+  cover?: string;
+  groupType?: number;
+}
