+++ src/components/Group/GroupModals/ExitGroupModal/index.module.less
@@ -1,0 +1,5 @@
+.exitGroupName {
+  margin: var(--ant-margin-lg) var(--ant-margin-sm);
+  font-weight: var(--ant-font-weight-md);
+  font-size: var(--ant-font-size-lg);
+}
