+++ src/components/Group/GroupModals/ExitGroupModal/index.type.ts
@@ -1,0 +1,7 @@
+export interface ExitGroupModalProps {
+  open: boolean;
+  onCancel: () => void;
+  onSuccess?: () => void;
+  groupName: string;
+  groupId?: string;
+}
