+++ src/components/Group/GroupModals/JoinGroupModal/index.type.ts
@@ -1,0 +1,5 @@
+export interface JoinGroupModalProps {
+  open: boolean;
+  onCancel: () => void;
+  onSuccess?: () => void;
+}
