+++ src/components/Group/GroupModals/index.ts
@@ -3,3 +3,9 @@ export { default as JoinGroupModal } from './JoinGroupModal';
 export { default as ExitGroupModal } from './ExitGroupModal';
 export { default as DissolveGroupModal } from './DissolveGroupModal';
 export { default as EditGroupInfoModal } from './EditGroupInfoModal';
+
+export type { CreateGroupModalProps } from './CreateGroupModal/index.type';
+export type { JoinGroupModalProps } from './JoinGroupModal/index.type';
+export type { ExitGroupModalProps } from './ExitGroupModal/index.type';
+export type { DissolveGroupModalProps } from './DissolveGroupModal/index.type';
+export type { EditGroupInfoModalProps } from './EditGroupInfoModal/index.type';
