--- src/components/Group/GroupModals/style.module.less
@@ -1,25 +1,0 @@
-.modalFormPadding {
-  padding: var(--ant-padding-lg) 0;
-}
-
-.fullWidth {
-  width: 100%;
-}
-
-.modalSection {
-  margin-top: var(--ant-margin-sm);
-}
-
-.modalSectionLabel {
-  margin-bottom: var(--ant-margin-xs);
-}
-
-.modalSectionSubLabel {
-  color: var(--ant-color-text-secondary);
-}
-
-.exitGroupName {
-  margin: var(--ant-margin-lg) var(--ant-margin-sm);
-  font-weight: var(--ant-font-weight-md);
-  font-size: var(--ant-font-size-lg);
-}
